Shared lib

+++ shared/math.c
@@ -1,5 +1,4 @@
-#include "../pch.h"
-
+#include "math.h"
 int pow(int n, int power) {
     int result = 1;
 
@@ -26,7 +25,7 @@ int min(int x, int y) {
 }
 int ceil(int x, int y) { return (x + y - 1) / y; }
 
-float abs(int x) {
+int abs(int x) {
     if (x < 0)
         return x * -1;
